move node testing events (W) to core events
