+++ core/dbt/common/events/types.proto
@@ -76,67 +76,6 @@ message RecordRetryExceptionMsg {
     RecordRetryException data = 2;
 }
 
-// W - Node testing
-
-// Skipped W001
-
-// W002
-message CatchableExceptionOnRun {
-    NodeInfo node_info = 1;
-    string exc = 2;
-    string exc_info = 3;
-}
-
-message CatchableExceptionOnRunMsg {
-    EventInfo info = 1;
-    CatchableExceptionOnRun data = 2;
-}
-
-// W003
-message InternalErrorOnRun {
-    string build_path = 1;
-    string exc = 2;
-}
-
-message InternalErrorOnRunMsg {
-    EventInfo info = 1;
-    InternalErrorOnRun data = 2;
-}
-
-// W004
-message GenericExceptionOnRun {
-    string build_path = 1;
-    string unique_id = 2;
-    string exc = 3;
-}
-
-message GenericExceptionOnRunMsg {
-    EventInfo info = 1;
-    GenericExceptionOnRun data = 2;
-}
-
-// W005
-message NodeConnectionReleaseError {
-    string node_name = 1;
-    string exc = 2;
-    string exc_info = 3;
-}
-
-message NodeConnectionReleaseErrorMsg {
-    EventInfo info = 1;
-    NodeConnectionReleaseError data = 2;
-}
-
-// W006
-message FoundStats {
-    string stat_line = 1;
-}
-
-message FoundStatsMsg {
-    EventInfo info = 1;
-    FoundStats data = 2;
-}
-
 // Z - Misc
 
 // Z001
