+++ core/dbt/task/base.py
@@ -27,10 +27,6 @@ from dbt.contracts.results import (
 from dbt.common.events.contextvars import get_node_info
 from dbt.common.events.functions import fire_event
 from dbt.common.events.types import (
-    CatchableExceptionOnRun,
-    InternalErrorOnRun,
-    GenericExceptionOnRun,
-    NodeConnectionReleaseError,
     LogDebugStackTrace,
     LogSkipBecauseError,
 )
@@ -38,6 +34,10 @@ from dbt.events.types import (
     SkippingDetails,
     NodeCompiling,
     NodeExecuting,
+    CatchableExceptionOnRun,
+    InternalErrorOnRun,
+    GenericExceptionOnRun,
+    NodeConnectionReleaseError,
 )
 from dbt.common.exceptions import (
     DbtRuntimeError,
